requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.1
bcrypt==4.1.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.12
django-filter==25.2
djangorestframework==3.17.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.11
joblib==1.5.2
numpy==1.26.2
pillow==12.2.0
psycopg2-binary==2.9.11
pydantic==2.5.0
pydantic_core==2.14.1
pymongo==4.6.0
python-dotenv==1.2.2
python-multipart==0.0.6
requests==2.32.5
scikit-learn==1.3.2
scipy==1.16.3
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.27.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
